Day 16 part 2
Janky assumption that doing them in serial provides the same answer as doing them in parallel.

Janky because one might think that paths change if they coordinate they can find better paths.
The result is better than the sum of the components. However, since the path of the components.
The thing is that it's not the case. The path of the human is already optimized for maximum flow.
So we can just process the valves that are left and try and find a path for the elephant and sum them.
Jank but works good enough.